Organize group form code into folder
