+++ packages/app/screens/feed/create group form/CreateGroupStack.js
@@ -1,11 +1,9 @@
-import { 
-  InitialCreateGroupScreen,
-  CreateGroupFormScreen1,
-  CreateGroupFormScreen2,
-  CreateGroupFormScreen3
-} from './CreateGroupForm';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { GroupProvider } from '@app/utils/GroupContext';
+import InitialCreateGroupScreen from './InitialCreateGroupScreen';
+import CreateGroupFormScreen1 from './CreateGroupFormScreen1';
+import CreateGroupFormScreen2 from './CreateGroupFormScreen2';
+import CreateGroupFormScreen3 from './CreateGroupFormScreen3';
 
 const Stack = createNativeStackNavigator();
